decoder: CF1 stream support 34bit [1/1]
PD#SWPL-177304

Problem:
decoder support stream 34bit

Solution:
config stream prefix

Verify:
s6_s905x5_bl201

+++ drivers/amvdec_ports/decoder/vdec_hevc_if.c
@@ -358,7 +358,7 @@ static int vdec_hevc_probe(unsigned long h_vdec,
 
 	if (ctx->stream_mode) {
 		aml_es_write(ctx, bs->dbuf, bs->addr, size, bs->timestamp);
-		vdec_write_stream_data(adapt_vdec, (u32)bs->addr, size);
+		vdec_write_stream_data(adapt_vdec, bs->addr, size);
 		return 0;
 	}
 
@@ -476,7 +476,7 @@ static int vdec_hevc_decode(unsigned long h_vdec,
 
 	if (ctx->stream_mode) {
 		aml_es_write(ctx, bs->dbuf, bs->addr, size, bs->timestamp);
-		vdec_write_stream_data(vdec, (u32)bs->addr, size);
+		vdec_write_stream_data(vdec, bs->addr, size);
 		return size;
 	}
